feat(packages.yaml): add 'jandedobbeleer/oh-my-posh', 'iterm2', and 'font-fira-code-nerd-font' to packages list feat(aliases_all.tmpl): add 'llt' alias for 'ls -lart --color' feat(dot_zshrc.tmpl): add conditional initialization for 'oh-my-posh'
The 'jandedobbeleer/oh-my-posh', 'iterm2', and 'font-fira-code-nerd-font' were added to the packages list to enhance the terminal experience with better visuals and functionality. The 'llt' alias was added to provide a quick and color-coded way to list files in the terminal. The 'oh-my-posh' initialization was added to the zshrc template to ensure it's properly set up if it's installed.