style(dot_p10k.zsh.tmpl, dot_vim/vimrc, dot_zshenv.tmpl, dot_zshrc.tmpl): add vim modeline and newline at EOF for better file formatting feat(dot_vim/vimrc): add shiftwidth setting for consistent indentation
The vim modeline has been added to the top of the .zsh and .tmpl files to ensure that vim recognizes the file type and applies the correct syntax highlighting and indentation rules. A newline has been added at the end of each file to adhere to the POSIX standard, which defines a line as a sequence of zero or more non-newline characters followed by a newline character.

In the vimrc file, the shiftwidth setting has been added to ensure consistent indentation when using the shift operators or auto-indentation. This makes the code easier to read and maintain.