feat(packages.yaml): add HashiCorp tap and Terraform to package list
Add the 'hashicorp/tap' repository and 'hashicorp/tap/terraform'
package to the list of packages. This change allows for the
installation and management of Terraform, a widely-used
infrastructure as code tool, through the package manager.
Including HashiCorp's tap ensures access to the latest versions
and updates directly from the source.