feat(nvim): enhance lazy setup and telescope plugin configuration
Add support for importing LSP plugins in the lazy setup to improve
modularity and organization of the Neovim configuration. Enable the
checker with notifications disabled to ensure plugins are up-to-date
without interrupting the user. Add the "folke/todo-comments.nvim"
plugin to the telescope configuration to allow users to search for
TODO comments efficiently. Introduce a new key mapping for finding
TODOs, enhancing productivity by providing quick access to pending
tasks within the codebase.