feat(packages.yaml): add neovim to the list of packages for installation
Neovim is added to the package list to enhance the development environment with a modern and extensible text editor, providing better support for plugins and configurations compared to traditional Vim.

chore(direnvrc): remove custom direnvrc configuration for poetry

The custom direnvrc script for managing Python virtual environments with Poetry is removed. This change is made to simplify the configuration and rely on standard practices or other tools for managing virtual environments, reducing maintenance overhead and potential conflicts.